requirements.txt
fastapi>=0.110.0,<0.111.0
uvicorn>=0.28.0,<0.29.0
langchain-openai>=0.1.0,<0.2.0
python-dotenv>=1.0.0,<2.0.0
crewai==0.28.8 # Pinning to a very stable version
crewai-tools==0.1.7 # Pinning to a compatible stable version
pydantic==2.6.1 # UPDATED: To satisfy crewai's dependency (>=2.4.2, <3.0.0)
langchain-community>=0.0.30,<0.0.31 # Specific version for PyPDFLoader
pypdf>=4.0.0,<5.0.0 # Explicitly needed for PDF processing
python-multipart # Add this line